Add files via upload
added green electricity as a toggle -- user input TRUE or FALSE under the "Sourcing" subheading in the Red "Process Operation" table. 
users can also define what the prices and emissions are for both conventional and green electricity, and then the TRUE/FALSE selects between the two. 
the current price for green electricity is 20% higher than that of conventional electricity (purely an assumption at the moment)